removed unnecesery dispose
