+++ lib/screens/buy/business.dart
@@ -1,13 +1,12 @@
 import 'dart:async';
 import 'dart:core';
 import 'package:flutter/material.dart';
-import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:fusecash/generated/i18n.dart';
 import 'package:fusecash/models/business.dart';
 import 'package:fusecash/models/token.dart';
 import 'package:fusecash/screens/send/send_amount_arguments.dart';
-import 'package:fusecash/utils/forks.dart';
+import 'package:fusecash/utils/transaction_row.dart';
 import 'package:fusecash/widgets/bottombar.dart';
 import 'package:fusecash/widgets/drawer.dart';
 import 'package:url_launcher/url_launcher.dart';
@@ -39,30 +38,6 @@ class BusinessPage extends StatefulWidget {
   _BusinessPageState createState() => _BusinessPageState();
 }
 
-String getCoverPhotoUrl(business, communityAddress) {
-  if (business.metadata.coverPhoto == null ||
-        business.metadata.coverPhoto == '') {
-         return 'https://cdn3.iconfinder.com/data/icons/abstract-1/512/no_image-512.png';
-  } else if (isPaywise(communityAddress)) {
-    return business.metadata.coverPhoto;
-  }
-  else {
-    return DotEnv().env['IPFS_BASE_URL'] + '/image/' + business.metadata.coverPhoto;
-  }
-}
-
-String getImageUrl(business, communityAddress) {
-  if (business.metadata.image == null ||
-        business.metadata.image == '') {
-         return 'https://cdn3.iconfinder.com/data/icons/abstract-1/512/no_image-512.png';
-  } else if (isPaywise(communityAddress)) {
-    return business.metadata.image;
-  }
-  else {
-    return DotEnv().env['IPFS_BASE_URL'] + '/image/' + business.metadata.image;
-  }
-}
-
 class _BusinessPageState extends State<BusinessPage> {
   GlobalKey<ScaffoldState> scaffoldState;
   Completer<GoogleMapController> _controller = Completer();
@@ -74,11 +49,6 @@ class _BusinessPageState extends State<BusinessPage> {
   @override
   void initState() {
     super.initState();
-  }
-
-  @override
-  void dispose() {
-    super.dispose();
   }
 
   @override
