+++ lib/screens/cash_home/cash_transactions.dart
@@ -3,9 +3,8 @@ import 'package:fusecash/generated/i18n.dart';
 import 'package:fusecash/models/transfer.dart';
 import 'dart:core';
 import 'package:fusecash/models/views/cash_wallet.dart';
-import 'package:contacts_service/contacts_service.dart';
 import 'package:fusecash/screens/cash_home/transaction_row.dart';
-import 'package:fusecash/utils/phone.dart';
+import 'package:fusecash/utils/transaction_row.dart';
 
 class CashTransactios extends StatefulWidget {
   CashTransactios({@required this.viewModel});
@@ -15,58 +14,12 @@ class CashTransactios extends StatefulWidget {
   createState() => new CashTransactionsState();
 }
 
-String deduceSign(Transfer transfer) {
-  if (transfer.type == 'SEND') {
-    return '-';
-  } else {
-    return '+';
-  }
-}
-
-Contact getContact(Transfer transfer, Map<String, String> reverseContacts,
-    List<Contact> contacts, String countryCode) {
-  String accountAddress = transfer.type == 'SEND' ? transfer.to : transfer.from;
-  if (accountAddress == null) {
-    return null;
-  }
-  if (reverseContacts.containsKey(accountAddress.toLowerCase())) {
-    String phoneNumber = reverseContacts[accountAddress.toLowerCase()];
-    if (contacts == null) return null;
-    for (Contact contact in contacts) {
-      for (Item contactPhoneNumber in contact.phones.toList()) {
-        if (formatPhoneNumber(contactPhoneNumber.value, countryCode) ==
-            phoneNumber) {
-          return contact;
-        }
-      }
-    }
-  }
-  return null;
-}
-
-Color deduceColor(Transfer transfer) {
-  if (transfer.isFailed()) {
-    return Color(0xFFE0E0E0);
-  } else {
-    if (transfer.type == 'SEND') {
-      return Color(0xFFFF0000);
-    } else {
-      return Color(0xFF00BE66);
-    }
-  }
-}
-
 class CashTransactionsState extends State<CashTransactios> {
   CashTransactionsState();
 
   @override
   void initState() {
     super.initState();
-  }
-
-  @override
-  void dispose() {
-    super.dispose();
   }
 
   renderTrasfers() {
