+++ lib/screens/cash_home/prize.dart
@@ -80,11 +80,6 @@ class _PrizeScreenState extends State<PrizeScreen> {
   }
 
   @override
-  void dispose() {
-    super.dispose();
-  }
-
-  @override
   Widget build(BuildContext _context) {
     return new StoreConnector<AppState, PrizeViewModel>(
         converter: PrizeViewModel.fromStore,
