+++ lib/screens/cash_home/transaction_row.dart
@@ -1,80 +1,15 @@
 import 'package:contacts_service/contacts_service.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:fusecash/generated/i18n.dart';
-import 'package:fusecash/models/business.dart';
 import 'package:fusecash/models/transaction.dart';
 import 'package:fusecash/models/transfer.dart';
 import 'package:fusecash/models/views/cash_wallet.dart';
-import 'package:fusecash/screens/buy/business.dart';
 import 'package:fusecash/screens/cash_home/cash_home.dart';
-import 'package:fusecash/screens/cash_home/cash_transactions.dart';
+import 'package:fusecash/utils/transaction_row.dart';
 import 'package:fusecash/screens/cash_home/dai_explained.dart';
 import 'package:fusecash/screens/cash_home/transaction_details.dart';
 import 'package:fusecash/utils/format.dart';
-
-String deducePhoneNumber(Transfer transfer, Map<String, String> reverseContacts,
-    {bool format = true, List<Business> businesses}) {
-  String accountAddress = transfer.type == 'SEND' ? transfer.to : transfer.from;
-  if (businesses != null && businesses.isNotEmpty) {
-    Business business = businesses.firstWhere(
-        (business) => business.account == accountAddress,
-        orElse: () => null);
-    if (business != null) {
-      return business.name;
-    }
-  }
-  if (reverseContacts.containsKey(accountAddress)) {
-    return reverseContacts[accountAddress];
-  }
-  if (format) {
-    return formatAddress(accountAddress);
-  } else {
-    return accountAddress;
-  }
-}
-
-dynamic getImage(Transfer transfer, Contact contact, CashWalletViewModel vm) {
-  if (transfer.isJoinCommunity() &&
-      vm.community.metadata.image != null &&
-      vm.community.metadata.image != '') {
-    return new NetworkImage(DotEnv().env['IPFS_BASE_URL'] +
-        '/image/' +
-        vm.community.metadata.image);
-  } else if (transfer.isGenerateWallet()) {
-    return new AssetImage(
-      'assets/images/generate_wallet.png',
-    );
-  } else if (transfer.isJoinBonus()) {
-    return new AssetImage(
-      'assets/images/join.png',
-    );
-  } else if (contact?.avatar != null) {
-    return new MemoryImage(contact.avatar);
-  }
-
-  String accountAddress = transfer.type == 'SEND' ? transfer.to : transfer.from;
-  Business business = vm.businesses.firstWhere(
-      (business) => business.account == accountAddress,
-      orElse: () => null);
-  if (business != null) {
-    return NetworkImage(getImageUrl(business, vm.communityAddress));
-  }
-  return new AssetImage('assets/images/anom.png');
-}
-
-Color deduceColor(Transfer transfer) {
-  if (transfer.isFailed()) {
-    return Color(0xFFE0E0E0);
-  } else {
-    if (transfer.type == 'SEND') {
-      return Color(0xFFFF0000);
-    } else {
-      return Color(0xFF00BE66);
-    }
-  }
-}
 
 class TransactionListItem extends StatelessWidget {
   final Transaction _transaction;
