+++ lib/screens/send/alpabet_list_scroll_view.dart
@@ -435,11 +435,6 @@ class _AlphabetListScollViewState extends State<_AlphabetListScollView> {
     super.initState();
   }
 
-  @override
-  void dispose() {
-    super.dispose();
-  }
-
   List<Widget> aToZ() {
     List<Widget> charList = [];
