+++ lib/screens/send/send_success.dart
@@ -24,12 +24,6 @@ class _SendSuccessScreenState extends State<SendSuccessScreen>
   }
 
   @override
-  void dispose() {
-    super.dispose();
-  }
-
-
-  @override
   Widget build(BuildContext context) {
     final SendAmountArguments args = ModalRoute.of(context).settings.arguments;
