+++ lib/screens/signup/signup.dart
@@ -28,12 +28,7 @@ class _SignupScreenState extends State<SignupScreen> {
   @override
   void initState() {
     super.initState();
-  }
-
-  @override
-  void dispose() {
-    super.dispose();
-  }
+  }  
 
   @override
   Widget build(BuildContext context) {
