+++ lib/screens/signup/verify.dart
@@ -22,11 +22,6 @@ class _VerifyScreenState extends State<VerifyScreen> {
   }
 
   @override
-  void dispose() {
-    super.dispose();
-  }
-
-  @override
   Widget build(BuildContext context) {
     return new StoreConnector<AppState, OnboardViewModel>(
         distinct: true,
