+++ lib/screens/splash/create_wallet.dart
@@ -21,11 +21,6 @@ class _CreateWalletState extends State<CreateWallet> {
   }
 
   @override
-  void dispose() {
-    super.dispose();
-  }
-
-  @override
   Widget build(BuildContext context) {
     return new StoreConnector<AppState, SplashViewModel>(
         distinct: true,
