+++ lib/screens/splash/splash.dart
@@ -32,11 +32,6 @@ class _SplashScreenState extends State<SplashScreen> {
   }
 
   @override
-  void dispose() {
-    super.dispose();
-  }
-
-  @override
   void initState() {
     super.initState();
     _slideController = HouseController(demoUpdated: _update);
